small modification to cater to the package structure in the utils and add notebooks
